requirements.txt
# Core dependencies
deepgram-sdk==3.0.0
cartesia==2.0.5
fastapi==0.110.0
uvicorn==0.27.1
aiortc==1.5.0
python-dotenv==1.1.0
websockets==12.0

# Web framework dependencies
starlette==0.36.3
jinja2==3.1.3
python-multipart==0.0.9

# Audio/video processing
av==10.0.0
numpy==1.26.4
scipy==1.11.4

# HTTP client
aiohttp>=3.11.12

# Optional: OpenAI (if you want to add AI responses later)
# openai>=1.70.0

# Pipecat for audio pipeline
pipecat==0.0.1
